+++ src/index.ts
@@ -1,8 +1,8 @@
-import Alert from './components/Alert'
-import Button from './components/Button'
-import ButtonGroup from './components/ButtonGroup'
+import { Alert } from './components/Alert'
+import { Button } from './components/Button'
+import { ButtonGroup } from './components/ButtonGroup'
 import { Modal } from './components/Modal'
-import Spinner from './components/Spinner'
+import { Spinner } from './components/Spinner'
 import { TMode, TShape } from './typings/types'
 
 export {