+++ src/stories/Alert.stories.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import { Meta, Story } from '@storybook/react'
-import Alert, { AlertProps } from '../components/Alert'
+import { Alert, AlertProps } from '../components/Alert'
 
 export default {
-  title: 'Components/Alert',
+  title: 'Featured/Alert',
   component: Alert,
 } as Meta