+++ src/stories/Button.stories.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import { Meta, Story } from '@storybook/react'
-import Button, { ButtonProps } from '../components/Button'
+import { Button, ButtonProps } from '../components/Button'
 
 export default {
-  title: 'Components/Button',
+  title: 'Featured/Button',
   component: Button,
 } as Meta